+++ bricks/next-builder/src/preview-container/PreviewContainer.tsx
@@ -21,6 +21,7 @@ import type {
   PreviewMessageContainerProxyMethod,
   PreviewMessageContainerStartPreview,
   PreviewMessageContainerToggleInspecting,
+  PreviewMessageContainerUpdatePreviewUrl,
   PreviewMessageFromContainer,
   PreviewMessageToContainer,
   PreviewSettings,
@@ -80,6 +81,7 @@ export interface PreviewContainerRef {
   reload(): void;
   capture(): void;
   resize(): void;
+  updatePreviewUrl(url: string): void;
   manager: BuilderDataManager;
 }
 
@@ -334,6 +336,8 @@ export function LegacyPreviewContainer(
   const handleIframeLoad = useCallback(() => {
     loadedRef.current = true;
     const snippetData = getSnippetData(snippetGraphData);
+    setHoverOutlines([]);
+    setActiveOutlines([]);
     iframeRef.current.contentWindow.postMessage(
       {
         sender: "preview-container",
@@ -491,6 +495,18 @@ export function LegacyPreviewContainer(
       previewOrigin
     );
   }, []);
+
+  const updatePreviewUrl = (url: string): void => {
+    iframeRef.current.contentWindow.postMessage(
+      {
+        sender: "preview-container",
+        type: "update-preview-url",
+        url,
+      } as PreviewMessageContainerUpdatePreviewUrl,
+      previewOrigin
+    );
+  };
+
   useImperativeHandle(ref, () => ({
     refresh,
     reload,
@@ -498,6 +514,7 @@ export function LegacyPreviewContainer(
     resize,
     manager,
     excuteProxyMethod,
+    updatePreviewUrl,
   }));
 
   useEffect(() => {