+++ bricks/next-builder/src/shared/workbench/WorkbenchTree.tsx
@@ -9,7 +9,7 @@ import React, {
   useEffect,
 } from "react";
 import { Input } from "antd";
-import { isMatch, pick } from "lodash";
+import { get, isMatch, pick } from "lodash";
 import classNames from "classnames";
 import { SearchOutlined } from "@ant-design/icons";
 import { smartDisplayForEvaluableString } from "@next-core/brick-utils";
@@ -135,11 +135,11 @@ export function WorkbenchTree({
     status: dragStatusEnum;
   } => {
     const element = findDragParent(e.target as HTMLElement);
-    if (element === curElement) {
+    const id = element.dataset.uid;
+    if (element === curElement || id.includes(":")) {
       return;
     }
     const { top, bottom } = element.getBoundingClientRect();
-    const id = element.dataset.uid;
     let status: dragStatusEnum;
     const repair = allowDragToInside ? 5 : 10;
     if (e.clientY < top + repair) {
@@ -315,6 +315,7 @@ export interface TreeNodeProps {
   level: number;
   isFirst?: boolean;
   isLast?: boolean;
+  skipNotify?: boolean;
 }
 
 function TreeNode({
@@ -340,6 +341,7 @@ function TreeNode({
     contextMenuFactory,
     onNodeToggle,
     getCollapsedId,
+    skipNotify = true,
   } = useWorkbenchTreeContext();
   const {
     allow,
@@ -508,7 +510,16 @@ function TreeNode({
           onContextMenu={onContextMenu}
           noEmptyHref
           onClick={onClick}
-          {...pick(node.link, ["to", "href"])}
+          {...(skipNotify
+            ? { ...pick(node.link, ["to", "href"]) }
+            : {
+                to: {
+                  pathname: get(node.link, "to", "href"),
+                  state: {
+                    notify: false,
+                  },
+                },
+              })}
         >
           <span
             className={styles.nodeLabel}