+++ bricks/next-builder/src/workbench-tree/index.tsx
@@ -78,6 +78,9 @@ export class WorkbenchTreeElement extends UpdatingElement {
   @property({ type: String })
   nodeKey: string;
 
+  @property({ type: Boolean })
+  skipNotify: boolean;
+
   @event({ type: "action.click" })
   private _actionClickEvent: EventEmitter<ActionClickDetail>;
 
@@ -88,9 +91,17 @@ export class WorkbenchTreeElement extends UpdatingElement {
   @event({ type: "node.click" })
   private _nodeClickEvent: EventEmitter<unknown>;
 
-  private _nodeClickFactory = (node: WorkbenchNodeData) => () => {
-    this._nodeClickEvent.emit(node.data);
-  };
+  private _nodeClickFactory =
+    (node: WorkbenchNodeData) => (event: React.MouseEvent) => {
+      // Q: It's weird that we MUST stop propagation here.
+      // Or this listener will be called twice.
+      // This may be a known issue of React 16 with shadow DOM,
+      // which is hopefully fixed in react 17.
+      // And a potential workaround for react 16:
+      // https://github.com/facebook/react/issues/9242#issuecomment-534096832
+      event.stopPropagation();
+      this._nodeClickEvent.emit(node.data);
+    };
 
   @event({ type: "node.drop" })
   private _nodeDropEvent: EventEmitter<any>;
@@ -147,6 +158,7 @@ export class WorkbenchTreeElement extends UpdatingElement {
                 fixedActionsFor: this.fixedActionsFor,
                 nodeKey: this.nodeKey,
                 collapsible: this.collapsible,
+                skipNotify: this.skipNotify,
                 clickFactory: this._nodeClickFactory,
                 contextMenuFactory: this._contextMenuFactory,
                 matchNode: (node, lowerTrimmedQuery) =>